+++ docs/user-tutorial/benchmarks/micro-benchmarks.md
@@ -419,23 +419,10 @@ Test the performance of distributed model inference.
 
 #### Metrics
 
-| Name                                            | Unit      | Description                                           |
-|-------------------------------------------------|-----------|-------------------------------------------------------|
-| pytorch-dist-inference/step_times               | time (ms) | Average time of model inference runs.                 |
-| pytorch-dist-inference/step_times_${percentile} | time (ms) | Tail (50,90,95,99,99.9) time of model inference runs. |
-
-### `dist-inference-cpp`
-
-#### Introduction
-
-Test the performance of distributed model inference (cpp implementation).
-
-#### Metrics
-
-| Name                                                  | Unit      | Description                                                        |
-|-------------------------------------------------------|-----------|--------------------------------------------------------------------|
-| dist-inference-cpp/layer_latency_${mnk}               | time (us) | Average per-layer latency of model inference runs.                 |
-| dist-inference-cpp/layer_latency_${mnk}_${percentile} | time (us) | Tail (50,90,95,99,99.9) per-layer latency of model inference runs. |
+| Name                                    | Unit      | Description                                           |
+|-----------------------------------------|-----------|-------------------------------------------------------|
+| dist-inference/step_times               | time (ms) | Average time of model inference runs.                 |
+| dist-inference/step_times_${percentile} | time (ms) | Tail (50,90,95,99,99.9) time of model inference runs. |
 
 ## Storage Benchmarks
