+++ superbench/config/azure_ndmv4.yaml
@@ -179,12 +179,6 @@ superbench:
       <<: *default_pytorch_mode
     dist-inference:
       <<: *dist_inference_pytorch_mode
-    dist-inference-cpp:
-      enable: true
-      modes:
-        - name: mpi
-          proc_num: 8
-          node_num: 1
     ib-traffic:
       enable: false
       modes:
